requirements/base.txt
django==5.2
djangorestframework==3.15.2
python-dotenv==1.0.1
drf-spectacular==0.27.2
